tests: merge both mock PyUSB devices
