+++ tests/_testutils.py
@@ -88,6 +88,42 @@ class MockHidapiDevice:
         return self.write(data)
 
 
+class MockPyusbDevice():
+    def __init__(self, vendor_id=None, product_id=None, release_number=None,
+                 serial_number=None, bus=None, address=None, port=None):
+        self.vendor_id = vendor_id
+        self.product_id = product_id
+        self.release_numer = release_number
+        self.serial_number = serial_number
+        self.bus = bus
+        self.address = address
+        self.port = port
+
+        self.open = noop
+        self.claim = noop
+        self.release = noop
+        self.close = noop
+
+        self._reset_sent()
+
+    def read(self, endpoint, length, timeout=None):
+        if len(self._responses):
+            return self._responses.popleft()
+        return [0] * length
+
+    def write(self, endpoint, data, timeout=None):
+        self._sent_xfers.append(('write', endpoint, data))
+
+    def ctrl_transfer(self, bmRequestType, bRequest, wValue=0, wIndex=0,
+                      data_or_wLength=None, timeout=None):
+        self._sent_xfers.append(('ctrl_transfer', bmRequestType, bRequest,
+                                 wValue, wIndex, data_or_wLength))
+
+    def _reset_sent(self):
+        self._sent_xfers = deque()
+        self._responses = deque()
+
+
 VirtualEeprom = namedtuple('VirtualEeprom', ['name', 'data'])
